+++ src/views/lists/List.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useCallback } from 'react'
 import { useParams, useHistory } from 'react-router-dom'
-import { BreadcrumbHeader, DropdownMore, Loading, Modal, NoItems } from '../../reusable'
+import { BreadcrumbHeader, Loading, Modal, NoItems, PrincipalButtons } from '../../reusable'
 import { LinkComponent, ListInfo } from "../../components/"
 import LinkCreate from '../links/LinkCreate'
 import ListEdit from './ListEdit'
@@ -10,7 +10,7 @@ import {
   CCard,
   CCardHeader,
   CCol,
-  CRow
+  CRow,
 } from '@coreui/react'
 
 export default function List() {
@@ -49,12 +49,14 @@ export default function List() {
   }, [fetchList])
 
   async function handleDelete() {
-    const { error } = await supabase
-      .from('lists')
-      .delete()
-      .eq('id', id)
-    if (error) console.log("error", error);
-    else history.push('/lists');
+    if (window.confirm('Tem certeza que você deseja excluir?')) {
+      const { error } = await supabase
+        .from('lists')
+        .delete()
+        .eq('id', id)
+      if (error) console.log("error", error);
+      else history.push('/lists');
+    }
   }
 
   if (loading) return (<Loading />)
@@ -64,14 +66,8 @@ export default function List() {
       <Modal show={modal} onClose={toogleModal} component={<ListEdit list={list} edit={list => setList(list)} />} />
       <CCol xs="12" sm="9" md="9">
         <CCard>
-          <CCardHeader color="secondary">
+          <CCardHeader color="secondary" className='text-break text-justify'>
             {list.name}
-            <div className="card-header-actions">
-              <DropdownMore
-                editAction={() => toogleModal()}
-                deleteAction={() => handleDelete(list.id)}
-              />
-            </div>
           </CCardHeader>
         </CCard>
         <BreadcrumbHeader title='Links' quantidade={links.length} component={<LinkCreate />} />
@@ -80,7 +76,11 @@ export default function List() {
         }
       </CCol>
       <CCol xs="12" sm="3" md="3">
-        <ListInfo list={list} />
+        <ListInfo list={list}>
+          <div className="card-header-actions">
+            <PrincipalButtons editAction={() => toogleModal()} deleteAction={() => handleDelete(list.id)} />
+          </div>
+        </ListInfo>
       </CCol>
     </CRow>
   )
